feat: 工程化适配与并发修复
- go.mod: 升级依赖版本
- 全局 API 增加 args ...ArgsForBaseAtomos 变参透传（CosmosGetElementID/SpawnAtom 等）
- 新增 base_atomos_helper_test 辅助测试
- 修复 element_local spawnLockMap 的 Remove/Unlock 顺序竞态
- 修复 base_atomos BaseAtomosHalt 状态字符串与 Stopping 冲突
- 修复 atom_local/cosmos_local/element_local 的类型断言逻辑

+++ app_logging.go
@@ -14,7 +14,7 @@ import (
 
 const (
 	// AppLoggingDefaultMaxSize Default log max size is 10MB.
-	AppLoggingDefaultMaxSize = 10 * 1024 * 1024
+	AppLoggingDefaultMaxSize = int64(10 * 1024 * 1024)
 	// AppLoggingAutoCleanupRatio Default auto cleanup size is 300MB.
 	AppLoggingAutoCleanupRatio = int64(30)
 	AppLoggingAutoCleanupOff   = -1
